Removes coupling to Fpl.Worker from Slackbot
