+++ src/Fpl.Workers/Models/Mappers/LiveEventsExtractor.cs
@@ -48,58 +48,20 @@ namespace FplBot.Core.Helpers
             }).WhereNotNull();
         }
 
-        public static IEnumerable<FinishedFixture> GetProvisionalFinishedFixtures(ICollection<Fixture> latestFixtures, ICollection<Fixture> current, ICollection<Team> teams, ICollection<Player> players)
+        public static IEnumerable<int> GetProvisionalFinishedFixtures(ICollection<Fixture> latestFixtures, ICollection<Fixture> current, ICollection<Team> teams, ICollection<Player> players)
         {
             if(latestFixtures == null)
-                return new List<FinishedFixture>();
+                return new List<int>();
 
             if (current == null)
-                return new List<FinishedFixture>();
+                return new List<int>();
 
             var latestFinished = latestFixtures.Where(f => f.FinishedProvisional);
             var currentFinished = current.Where(f => f.FinishedProvisional);
             var newFinished = latestFinished.Except(currentFinished, new FixtureComparer());
             if (newFinished.Any())
-                return newFinished.Select(n => CreateFinishedFixture(teams, players, n));
-            return new List<FinishedFixture>();
-        }
-
-        public static FinishedFixture CreateFinishedFixture(ICollection<Team> teams, ICollection<Player> players, Fixture n)
-        {
-            return new FinishedFixture
-            {
-                Fixture = n,
-                HomeTeam = teams.First(t => t.Id == n.HomeTeamId),
-                AwayTeam = teams.First(t => t.Id == n.AwayTeamId),
-                BonusPoints = CreateBonusPlayers(players, n)
-            };
-        }
-
-        public static IEnumerable<BonusPointsPlayer> CreateBonusPlayers(ICollection<Player> players, Fixture fixture)
-        {
-            try
-            {
-                var bonusPointsHome = fixture.Stats.FirstOrDefault(s => s.Identifier == "bps")?.HomeStats;
-                var bonusPointsAway = fixture.Stats.FirstOrDefault(s => s.Identifier == "bps")?.AwayStats;
-
-                var home = bonusPointsHome.Select(BpsFilter).ToList();
-                var away = bonusPointsAway.Select(BpsFilter).ToList();
-                var aggregated = home.Concat(away).OrderByDescending(bpp => bpp.BonusPoints);
-                return aggregated;
-
-                BonusPointsPlayer BpsFilter(FixtureStatValue bps)
-                {
-                    return new BonusPointsPlayer
-                    {
-                        Player = players.First(p => p.Id == bps.Element),
-                        BonusPoints = bps.Value
-                    };
-                }
-            }
-            catch
-            {
-                return new List<BonusPointsPlayer>();
-            }
+                return newFinished.Select(n => n.Id);
+            return new List<int>();
         }
     }
