+++ src/FplBot.Core/FplBot.Core.csproj
@@ -17,7 +17,6 @@
     <ProjectReference Include="..\Fpl.Client\Fpl.Client.csproj" />
     <ProjectReference Include="..\ext\Slackbot.Net.SlackClients.Http\Slackbot.Net.SlackClients.Http.csproj" />
     <ProjectReference Include="..\ext\Slackbot.Net.Endpoints\Slackbot.Net.Endpoints.csproj" />
-    <ProjectReference Include="..\Fpl.Workers\Fpl.Workers.csproj" />
     <ProjectReference Include="..\FplBot.Data\FplBot.Data.csproj" />
     <ProjectReference Include="..\FplBot.Messaging.Contracts\FplBot.Messaging.Contracts.csproj" />     
     <ProjectReference Include="..\Fpl.Search\Fpl.Search.csproj" />
