+++ src/FplBot.Core/Handlers/FplEvents/FixtureFulltimeHandler.cs
@@ -1,6 +1,8 @@
+using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 using Fpl.Client.Abstractions;
+using Fpl.Client.Models;
 using FplBot.Core.Extensions;
 using FplBot.Core.Helpers;
 using FplBot.Data.Abstractions;
@@ -38,7 +40,7 @@ namespace FplBot.Core.GameweekLifecycle.Handlers
             var settings = await _settingsClient.GetGlobalSettings();
             var fixtures = await _fixtureClient.GetFixtures();
             var fplfixture = fixtures.FirstOrDefault(f => f.Id == message.FixtureId);
-            var fixture = LiveEventsExtractor.CreateFinishedFixture(settings.Teams, settings.Players, fplfixture);
+            var fixture = CreateFinishedFixture(settings.Teams, settings.Players, fplfixture);
             var title = $"*FT: {fixture.HomeTeam.ShortName} {fixture.Fixture.HomeTeamScore}-{fixture.Fixture.AwayTeamScore} {fixture.AwayTeam.ShortName}*";
             var threadMessage = Formatter.FormatProvisionalFinished(fixture);
 
@@ -48,6 +50,44 @@ namespace FplBot.Core.GameweekLifecycle.Handlers
                 {
                     await context.SendLocal(new PublishFulltimeMessageToSlackWorkspace(slackTeam.TeamId, title, threadMessage));
                 }
+            }
+        }
+
+        private static FinishedFixture CreateFinishedFixture(ICollection<Team> teams, ICollection<Player> players, Fixture n)
+        {
+            return new FinishedFixture
+            {
+                Fixture = n,
+                HomeTeam = teams.First(t => t.Id == n.HomeTeamId),
+                AwayTeam = teams.First(t => t.Id == n.AwayTeamId),
+                BonusPoints = CreateBonusPlayers(players, n)
+            };
+        }
+
+        private static IEnumerable<BonusPointsPlayer> CreateBonusPlayers(ICollection<Player> players, Fixture fixture)
+        {
+            try
+            {
+                var bonusPointsHome = fixture.Stats.FirstOrDefault(s => s.Identifier == "bps")?.HomeStats;
+                var bonusPointsAway = fixture.Stats.FirstOrDefault(s => s.Identifier == "bps")?.AwayStats;
+
+                var home = bonusPointsHome.Select(BpsFilter).ToList();
+                var away = bonusPointsAway.Select(BpsFilter).ToList();
+                var aggregated = home.Concat(away).OrderByDescending(bpp => bpp.BonusPoints);
+                return aggregated;
+
+                BonusPointsPlayer BpsFilter(FixtureStatValue bps)
+                {
+                    return new BonusPointsPlayer
+                    {
+                        Player = players.First(p => p.Id == bps.Element),
+                        BonusPoints = bps.Value
+                    };
+                }
+            }
+            catch
+            {
+                return new List<BonusPointsPlayer>();
             }
         }
 
@@ -64,5 +104,20 @@ namespace FplBot.Core.GameweekLifecycle.Handlers
                 });
             }
         }
+    }
+
+    public class FinishedFixture
+    {
+        public Fixture Fixture { get; set; }
+        public Team HomeTeam { get; set; }
+        public Team AwayTeam { get; set; }
+
+        public IEnumerable<BonusPointsPlayer> BonusPoints { get; set; } = new List<BonusPointsPlayer>();
+    }
+
+    public class BonusPointsPlayer
+    {
+        public Player Player { get; set; }
+        public int BonusPoints { get; set; }
     }
 }
