+++ src/FplBot.Tests/FplBot.Tests.csproj
@@ -45,6 +45,7 @@
 
     <ItemGroup>
       <ProjectReference Include="..\FplBot.Core\FplBot.Core.csproj" />
+      <ProjectReference Include="..\Fpl.Workers\Fpl.Workers.csproj" />
     </ItemGroup>
 
 </Project>
